+++ utils/time.js
@@ -20,6 +20,72 @@ export let getTimeStamp = () => {
   return Date.now();
 };
 
+// @type (Date, Date, Array<string>) => string
+//
+// Takes two dates and returns a human readable duration such as "in 2 minutes" or "4 days ago".
+export let getRelDisplayDate = (date, refDate = new Date(), locales = ["en"]) => {
+  let seconds = 60;
+  let minutes = seconds * 60;
+  let hours = minutes * 24;
+  let days = hours * 7;
+  let weeks = days * 7;
+
+  let time = date.getTime();
+  let refTime = refDate.getTime();
+
+  let formatter = new Intl.RelativeTimeFormat(locales, {numeric: "auto"});
+  let diff = Math.round((refTime - time) / 1000);
+  let sign;
+
+  // Past
+  if (diff < 0) {
+    sign = 1;
+    diff = -diff;
+  }
+  // Future
+  else {
+    sign = -1;
+    diff = diff;
+  }
+
+  if (diff < seconds) {
+    return formatter.format((diff * sign), "seconds");
+  }
+  else if (diff < minutes) {
+    return formatter.format(Math.round((diff * sign) / seconds), "minutes");
+  }
+  else if (diff < hours) {
+    return formatter.format(Math.round((diff * sign) / minutes), "hours");
+  }
+  else if (diff < days) {
+    return formatter.format(Math.round((diff * sign) / hours), "days");
+  }
+  else if (diff < weeks) {
+    return formatter.format(Math.round((diff * sign) / days), "weeks");
+  }
+  else {
+    let months = (date.getFullYear() - refDate.getFullYear()) * 12;
+    months -= refDate.getMonth();
+    months += date.getMonth();
+
+    if (Math.abs(months) > 12) {
+      if (months < 0) {
+        return formatter.format(Math.ceil(months/12), "years");
+      }
+      else {
+        return formatter.format(Math.floor(months/12), "years");
+      }
+    }
+    else {
+      return formatter.format(months, "months");
+    }
+  }
+};
+
+
+
+
+
 // @type (Function, number, Object) => Function
 // @src [https://github.com/documentcloud/underscore/blob/master/underscore.js#L627]
 // @license MIT License [https://github.com/documentcloud/underscore/blob/master/LICENSE]