+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 0.17.0 (PENDING)
 
 - Add `Xel.localesIds` getter
+- Add `RELDATETIME` FTL function
 - [Bugfix] Invisible unicode characters added to <x-message> args
 
 ## 0.16.4 (2022-05-31)